+++ ipcrawler/default-plugins/dnsrecon-subdomain-bruteforce.py
@@ -17,6 +17,7 @@ class DnsReconSubdomainBruteforce(ServiceScan):
 		if which('dnsrecon') is None:
 			self.error('The program dnsrecon could not be found. Make sure it is installed. (On Kali, run: sudo apt install dnsrecon)')
 			return False
+		return True
 
 	def manual(self, service, plugin_was_run):
 		domain_name = '<DOMAIN-NAME>'