+++ ipcrawler/default-plugins/oracle-scanner.py
@@ -15,6 +15,7 @@ class OracleScanner(ServiceScan):
 		if which('oscanner') is None:
 			self.error('The oscanner program could not be found. Make sure it is installed. (On Kali, run: sudo apt install oscanner)')
 			return False
+		return True
 
 	async def run(self, service):
 		await service.execute('oscanner -v -s {address} -P {port} 2>&1', outfile='{protocol}_{port}_oracle_scanner.txt')