+++ ipcrawler/default-plugins/oracle-tnscmd.py
@@ -15,6 +15,7 @@ class OracleTNScmd(ServiceScan):
 		if which('tnscmd10g') is None:
 			self.error('The tnscmd10g program could not be found. Make sure it is installed. (On Kali, run: sudo apt install tnscmd10g)')
 			return False
+		return True
 
 	async def run(self, service):
 		if service.target.ipversion == 'IPv4':