+++ ipcrawler/default-plugins/redis-cli.py
@@ -15,6 +15,7 @@ class RedisCli(ServiceScan):
 		if which('redis-cli') is None:
 			self.error('The redis-cli program could not be found. Make sure it is installed. (On Kali, run: sudo apt install redis-tools)')
 			return False
+		return True
 
 	async def run(self, service):
 		if which('redis-cli') is not None: